update tick to be u32 instead of u16

+++ FluxPilot/crates/light_machine/src/lib.rs
@@ -295,7 +295,7 @@ impl<'a, 'b> Program<'a, 'b> {
         &mut self,
         machine_number: ProgramWord,
         index: u16,
-        tick: u16,
+        tick: u32,
         stack: &mut Vec<StackWord, STACK_SIZE>,
     ) -> Result<(u8, u8, u8), MachineError> {
         if stack.len() < 3 {
